Add support for reset / disconnect to server
The command logic doesn't need to update the status during the 'reset'
or 'disconnect' commands so that the state stays at disconnected.
Otherwise we're advancing the state to 'INITIALIZED'.

This might be OK, but we need to define the semantics of
DISCONNECTED+INITIALIZED versus just DISCONNECTED.  This probably
doesn't matter for the demo though?